requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
Pillow==10.0.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
